+++ commands/sense/command.js
@@ -1,50 +1,13 @@
-const buildLegacy = require('./lib/build-legacy');
-const build = require('./lib/build');
+const init = require('./lib/init-config');
+const sense = require('./lib/sense');
 
 module.exports = {
   command: 'sense',
   desc: 'Build a nebula visualization as a Qlik Sense extension',
   builder(yargs) {
-    yargs.option('ext', {
-      type: 'string',
-      required: false,
-      desc: 'Extension definition',
-    });
-    yargs.option('meta', {
-      type: 'string',
-      required: false,
-      desc: 'Extension meta information',
-    });
-    yargs.option('output', {
-      type: 'string',
-      required: false,
-      default: '<name>-ext',
-      desc: 'Destination directory',
-    });
-    yargs.option('minify', {
-      type: 'boolean',
-      required: false,
-      default: true,
-      desc: 'Minify and uglify code',
-    });
-    yargs.option('sourcemap', {
-      type: 'boolean',
-      required: false,
-      default: false,
-      desc: 'Generate sourcemaps',
-    });
-    yargs.option('legacy', {
-      type: 'boolean',
-      required: false,
-      default: false,
-      desc: 'Generate legacy extension',
-    });
+    init(yargs);
   },
   handler(argv) {
-    if (argv.legacy) {
-      buildLegacy(argv);
-    } else {
-      build(argv);
-    }
+    sense(argv);
   },
 };